Add files via upload
This project contains finance and risk analytics reports of companies.

Problem Statement - Businesses or companies can fall prey to default if they cannot keep up their debt obligations. Defaults will lead to a lower credit rating for the company which reduces its chances of getting credit in the future and may have to pay higher interest on existing debts and any new obligations. From an investor's point of view, he would want to invest in a company if it can handle its financial obligations, can grow quickly, and can manage the growth scale.
A balance sheet is a financial statement of a company that provides a snapshot of what a company owns, owes, and the amount invested by the shareholders. Thus, it is an important tool that helps evaluate the performance of a business.
Data that is available includes information from the financial statement of the companies for the previous year (2015). Also, information about the Net worth of the company in the following year (2016) is provided which can drive the labeled field.