Add files via upload
Election Exit Poll Prediction and U.S.A Presidential Speech Analysis using Machine Learning

Course Machine Learning

This project is based on 2 case studies: Vote Prediction and Text Analysis. The first project is to predict which party a citizen is going to vote for based on their age and according to the answers given by the citizens to the questions asked in a survey conducted. The second project is based on the analysis of the inaugural U.S.A. Presidential speeches. One has to draw inferences based on the analysis done on these speeches.

Skills and Tools

Text Mining Analytics, Support Vector Machine - K Nearest Neighbour - Naive Bayes, Ensemble Techniques, Logistic Regression - Linear Discriminant Analysis